Refined wsvimbufp & fixed a bug of console.
wsvimbufp supports opening an existing file.
THe bug that the console cannot finish a block which using a decorator
is fixed.

+++ wavesynlib/interfaces/os/gadgets/wsvimbufp.py
@@ -6,29 +6,88 @@ Created on Sun Mar 18 21:16:32 2018
 """
 
 import os
+import sys
+import getopt
+import chardet
 import tempfile
 import subprocess
 
 
 
+ERROR_NOERROR, ERROR_PARAM = range(2)
+
+
+
+def detect_file_encoding(filename):
+    detector = chardet.UniversalDetector()
+    with open(filename, 'rb') as f:
+        for line in f:
+            detector.feed(line)
+            if detector.done: break
+        detector.close()
+    return detector.result
+
+
 def main(argv):
-    fd, filename = tempfile.mkstemp(text=True)
-    with os.fdopen(fd, 'w') as f:
-        # Close the temp file and consequently the external editor can edit
-        # it without limitations.
-        pass
-    subprocess.call(['gvim', filename])
-    with open(filename, 'r') as f:
+    try:
+        opts, args = getopt.getopt(argv[1:],\
+            'gc:',\
+            [])
+    except getopt.GetoptError as err:
+        print(str(err), file=sys.stderr)
+        return ERROR_PARAM
+    
+    vim_name = 'vim'
+    
+    opt_list = []
+    for o, a in opts:
+        if o == '-g':
+            vim_name = 'gvim'
+        else:
+            opt_list.append(o)
+            if a:
+                opt_list.append(a)
+           
+    is_stdin = False
+    is_temp = True
+    
+    if args:
+        if args[0] != '-':
+            filename = args[0]
+            is_temp = False
+        else:
+            is_stdin = True
+            
+    if not is_temp:
+        filename = args[0]
+    else:
+        fd, filename = tempfile.mkstemp(text=True)
+        args.append(filename)        
+        with os.fdopen(fd, 'w') as f:
+            # Close the temp file and consequently the external editor can edit
+            # it without limitations.
+            if is_stdin:
+                f.write(sys.stdin.read())
+                
+    opt_list.append(filename)
+    subprocess.call([vim_name, *opt_list])
+    
+    detect_result = detect_file_encoding(filename)
+    # If multiple files are given, only output the content of the first file. 
+    with open(filename, 'r', encoding=detect_result['encoding']) as f:
         for line in f:
             print(line, end='')
-    os.remove(filename)
+    if is_temp:
+        os.remove(filename)
+    return ERROR_NOERROR
     
     
     
 if __name__ == '__main__':
-    main(None)
+    main(sys.argv)
 
 
 
 # Powershell Example:
-# python -c $(wsvimbufp | Out-String)
+# python -c $(wsvimbufp -g | Out-String)
+# using -g is highly recommanded on Windows OS.
